Rework Auth
