+++ src/routes/login/callback/+server.ts
@@ -1,9 +1,7 @@
 import { error, redirect } from '@sveltejs/kit';
-import { DISCORD_CLIENT_SECRET } from '$env/static/private';
-import { PUBLIC_DISCORD_REDIRECT_ROUTE, PUBLIC_DISCORD_CLIENT_ID } from '$env/static/public';
 import type { RequestHandler } from './$types';
-import type { components } from '$lib/api/api';
 import { LoginUser } from '$lib/api/elite';
+import { PUBLIC_DISCORD_REDIRECT_ROUTE } from '$env/static/public';
 
 export const GET: RequestHandler = async ({ url, cookies }) => {
 	const code = url.searchParams.get('code');
@@ -28,38 +26,13 @@ export const GET: RequestHandler = async ({ url, cookies }) => {
 
 	const [, redirectTo = '', attemptCount = 0] = state.split('|');
 
-	const data = {
-		client_id: PUBLIC_DISCORD_CLIENT_ID,
-		client_secret: DISCORD_CLIENT_SECRET,
-		grant_type: 'authorization_code',
-		code: code,
-		redirect_uri: url.origin + PUBLIC_DISCORD_REDIRECT_ROUTE,
-		// state: state,
-		// scope: 'identify guilds role_connections.write',
-	};
-
-	const request = await fetch('https://discord.com/api/v10/oauth2/token', {
-		method: 'POST',
-		body: new URLSearchParams(data),
-		headers: {
-			'Content-Type': 'application/x-www-form-urlencoded',
-		},
-	});
-
-	const response = (await request.json()) as components['schemas']['DiscordLoginDto'] & { error?: unknown };
-
-	if (response.error) {
-		throw error(400, new Error('Discord Authentication Error'));
-	}
-
 	const {
 		data: loginResponse,
 		response: r,
 		error: e,
 	} = await LoginUser({
-		access_token: response.access_token,
-		expires_in: Math.floor(+response.expires_in / 1000).toString(),
-		refresh_token: response.refresh_token,
+		code: code,
+		redirect_uri: url.origin + PUBLIC_DISCORD_REDIRECT_ROUTE,
 	});
 
 	if (!loginResponse) {
@@ -72,7 +45,7 @@ export const GET: RequestHandler = async ({ url, cookies }) => {
 	const refreshTokenExpires = new Date(Date.now() + thirtyDays * 1000); // 30 days
 
 	cookies.set('access_token', loginResponse.access_token, {
-		// The access token expires every 10 minutes, but we keep it to use with the refresh token
+		// The access token expires sooner, but we keep it to use with the refresh token
 		expires: refreshTokenExpires,
 		maxAge: thirtyDays,
 		path: '/',
