+++ src/routes/profile/servers/+page.server.ts
@@ -2,7 +2,7 @@ import { redirect } from '@sveltejs/kit';
 import type { PageServerLoad } from './$types';
 import { GetPublicGuilds, GetUsersGuilds } from '$lib/api/elite';
 import type { components } from '$lib/api/api';
-import { FetchDiscordUserData } from '$lib/discordAuth';
+import { FetchDiscordUserData } from '$lib/api/auth';
 
 export const load: PageServerLoad = async ({ locals, parent, url }) => {
 	const { session } = await parent();
